Move types used in mempool to mem_pool_types
